Snapshot outputs after task
