+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/AbstractNamedFileSnapshotTaskStateChanges.java
@@ -18,82 +18,35 @@ package org.gradle.api.internal.changedetection.rules;
 
 import com.google.common.base.Function;
 import com.google.common.collect.ImmutableSortedMap;
-import com.google.common.collect.ImmutableSortedSet;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Iterators;
 import com.google.common.collect.Lists;
-import org.gradle.api.UncheckedIOException;
+import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.internal.changedetection.state.FileCollectionSnapshot;
-import org.gradle.api.internal.changedetection.state.FileCollectionSnapshotter;
-import org.gradle.api.internal.changedetection.state.FileCollectionSnapshotterRegistry;
 import org.gradle.api.internal.changedetection.state.TaskExecution;
-import org.gradle.api.internal.tasks.TaskFilePropertySpec;
-import org.gradle.normalization.internal.InputNormalizationStrategy;
 import org.gradle.util.ChangeListener;
 import org.gradle.util.DiffUtil;
 
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.SortedSet;
 
 abstract public class AbstractNamedFileSnapshotTaskStateChanges implements TaskStateChanges {
-    private final String taskName;
-    private final String title;
-    private final ImmutableSortedSet<? extends TaskFilePropertySpec> fileProperties;
-    private final FileCollectionSnapshotterRegistry snapshotterRegistry;
     protected final TaskExecution previous;
     protected final TaskExecution current;
-    private final InputNormalizationStrategy normalizationStrategy;
-    protected ImmutableSortedMap<String, FileCollectionSnapshot> currentSnapshots;
+    private final TaskInternal task;
+    private final String title;
 
-    protected AbstractNamedFileSnapshotTaskStateChanges(String taskName, TaskExecution previous, TaskExecution current, FileCollectionSnapshotterRegistry snapshotterRegistry, String title, ImmutableSortedSet<? extends TaskFilePropertySpec> fileProperties, InputNormalizationStrategy normalizationStrategy, ImmutableSortedMap<String, FileCollectionSnapshot> currentSnapshots) {
-        this.taskName = taskName;
+    protected AbstractNamedFileSnapshotTaskStateChanges(TaskExecution previous, TaskExecution current, TaskInternal task, String title) {
         this.previous = previous;
         this.current = current;
-        this.snapshotterRegistry = snapshotterRegistry;
+        this.task = task;
         this.title = title;
-        this.fileProperties = fileProperties;
-        this.normalizationStrategy = normalizationStrategy;
-        this.currentSnapshots = currentSnapshots;
-    }
-
-    protected String getTaskName() {
-        return taskName;
-    }
-
-    protected String getTitle() {
-        return title;
-    }
-
-    protected ImmutableSortedSet<? extends TaskFilePropertySpec> getFileProperties() {
-        return fileProperties;
     }
 
     protected abstract ImmutableSortedMap<String, FileCollectionSnapshot> getPrevious();
 
-    protected FileCollectionSnapshotterRegistry getSnapshotterRegistry() {
-        return snapshotterRegistry;
-    }
-
-    protected ImmutableSortedMap<String, FileCollectionSnapshot> getCurrent() {
-        return currentSnapshots;
-    }
-
-    protected ImmutableSortedMap<String, FileCollectionSnapshot> buildSnapshots(String taskName, FileCollectionSnapshotterRegistry snapshotterRegistry, String title, SortedSet<? extends TaskFilePropertySpec> fileProperties) {
-        ImmutableSortedMap.Builder<String, FileCollectionSnapshot> builder = ImmutableSortedMap.naturalOrder();
-        for (TaskFilePropertySpec propertySpec : fileProperties) {
-            FileCollectionSnapshot result;
-            try {
-                FileCollectionSnapshotter snapshotter = snapshotterRegistry.getSnapshotter(propertySpec.getSnapshotter());
-                result = snapshotter.snapshot(propertySpec.getPropertyFiles(), propertySpec.getPathNormalizationStrategy(), normalizationStrategy);
-            } catch (UncheckedIOException e) {
-                throw new UncheckedIOException(String.format("Failed to capture snapshot of %s files for task '%s' property '%s' during up-to-date check.", title.toLowerCase(), taskName, propertySpec.getPropertyName()), e);
-            }
-            builder.put(propertySpec.getPropertyName(), result);
-        }
-        return builder.build();
-    }
+    protected abstract ImmutableSortedMap<String, FileCollectionSnapshot> getCurrent();
 
     protected Iterator<TaskStateChange> getFileChanges(final boolean includeAdded) {
         if (getPrevious() == null) {
@@ -103,12 +56,12 @@ abstract public class AbstractNamedFileSnapshotTaskStateChanges implements TaskS
         DiffUtil.diff(getCurrent().keySet(), getPrevious().keySet(), new ChangeListener<String>() {
             @Override
             public void added(String element) {
-                propertyChanges.add(new DescriptiveChange("%s property '%s' has been added for task '%s'", title, element, taskName));
+                propertyChanges.add(new DescriptiveChange("%s property '%s' has been added for %s", title, element, task));
             }
 
             @Override
             public void removed(String element) {
-                propertyChanges.add(new DescriptiveChange("%s property '%s' has been removed for task '%s'", title, element, taskName));
+                propertyChanges.add(new DescriptiveChange("%s property '%s' has been removed for %s", title, element, task));
             }
 
             @Override
