+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/InputFilesTaskStateChanges.java
@@ -19,16 +19,19 @@ package org.gradle.api.internal.changedetection.rules;
 import com.google.common.collect.ImmutableSortedMap;
 import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.internal.changedetection.state.FileCollectionSnapshot;
-import org.gradle.api.internal.changedetection.state.FileCollectionSnapshotterRegistry;
 import org.gradle.api.internal.changedetection.state.TaskExecution;
-import org.gradle.normalization.internal.InputNormalizationStrategy;
 
 import javax.annotation.Nullable;
 import java.util.Iterator;
 
 public class InputFilesTaskStateChanges extends AbstractNamedFileSnapshotTaskStateChanges {
-    public InputFilesTaskStateChanges(@Nullable TaskExecution previous, TaskExecution current, TaskInternal task, FileCollectionSnapshotterRegistry snapshotterRegistry, InputNormalizationStrategy normalizationStrategy) {
-        super(task.getName(), previous, current, snapshotterRegistry, "Input", task.getInputs().getFileProperties(), normalizationStrategy, current.getInputFilesSnapshot());
+    public InputFilesTaskStateChanges(@Nullable TaskExecution previous, TaskExecution current, TaskInternal task) {
+        super(previous, current, task, "Input");
+    }
+
+    @Override
+    protected ImmutableSortedMap<String, FileCollectionSnapshot> getCurrent() {
+        return current.getInputFilesSnapshot();
     }
 
     @Override
